+++ cvat-ui/src/components/annotation-page/standard-workspace/controls-side-bar/draw-cuboid-control.tsx
@@ -12,6 +12,7 @@ import { ShapeType } from 'reducers/interfaces';
 import { CubeIcon } from 'icons';
 
 import DrawShapePopoverContainer from 'containers/annotation-page/standard-workspace/controls-side-bar/draw-shape-popover';
+import withVisibilityHandling from './handle-popover-visibility';
 
 interface Props {
     canvasInstance: Canvas;
@@ -20,6 +21,7 @@ interface Props {
 
 function DrawPolygonControl(props: Props): JSX.Element {
     const { canvasInstance, isDrawing } = props;
+    const CustomPopover = withVisibilityHandling(Popover, 'draw-cuboid');
 
     const dynamcPopoverPros = isDrawing ?
         {
@@ -41,14 +43,14 @@ function DrawPolygonControl(props: Props): JSX.Element {
         };
 
     return (
-        <Popover
+        <CustomPopover
             {...dynamcPopoverPros}
             overlayClassName='cvat-draw-shape-popover'
             placement='right'
             content={<DrawShapePopoverContainer shapeType={ShapeType.CUBOID} />}
         >
             <Icon {...dynamicIconProps} component={CubeIcon} />
-        </Popover>
+        </CustomPopover>
     );
 }